bin/battery-info: Show information about the battery
The main purpose of this is to give you the calculated full
capacity of the particular battery attached to a laptop. On old
batteries this can be a small fraction of the design capacity
(also shown), indicating that the battery can't be relied upon to
last any significant length of time.

It also gives model and serial number.