bin/backlight: Check or set display backlight level
This works only on systems with integrated displays, e.g., laptops.

With no params, this will set it to 75% to make it easier to recover
from turning the backlight completely off.

(This was originally written because, though the "lower backlight" key
works on Thinkpads using Debian 7, the "raise backlight" key seems to
be broken.)