bin/genmon-vclock: Fix newline issues by converting to Bash
`/bin/sh` on Debian is dash, but on CentOS it's a symlink to Bash.
Unfortuantely, the dash built-in echo always interprets escape
sequences as if the `-e` option had been provided to the Bash echo.
This caused the obvious problems when running this script on CentOS
systems.

We fix this by just using Bash since it's always there and gives us
consistent behaviour. And remind ourselves never to use `/bin/sh`
again.