bin/show-terminal-colors: Add -l option to pipe output through less(1)
As well as being useful when more than a screenful of output is
produced, this is handy when running this as a command inside a new
terminal (e.g., with `xterm -e`). To handle this case properly we set
the less options so that there will always be a prompt before exiting,
avoiding the terminal popping up and immediately closing.