bin/show-terminal-colors: Add -t option to start new terminals
One or more -t PROG options will make this start new terminals and run
itself in them. The new -r option, to reload `~/.Xresources` can also
be handy when doing this as it allows you to tweak the file (typically
terminal colour and font settings) and quickly view the results of the
changes.

We currently hardcode the first terminal's starting location to 50
pixels away from the lower left-hand corner of the desktop; each
terminal after that will be immediately to the right of the previous
terminal.