bin/genmon-vclock: Left-justify everything
The width you want for the xfce4-panel, and thus the size of the font
you want to use, is dependent on your screen size. (A nice-looking
width varies from about 20 pixels on a 1024x768 monitor to 30 pixels
on a 4k monitor.) After choosing a width in the xfce4-panel
preferences you can then set an appropriately sized font in the genmon
properties for the monitor showing this clock. (Note that the font
won't update until you click the "Close" button in the dialogue after
setting the font.)

We left-justify everything to give maximum flexibility in choosing
font size; this allows choosing a larger font that cuts off some of
the right-hand data, such as the third letter of the weekday or time
zone.